araba-iq-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]>=3.2.4,<4
pydantic-settings==2.6.1
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.49.1
